// src/app/models/paging.ts
export class Paging {
    CategoryId: number;
    SubCategoryId: number;
    CategoryName: string;
    SubCategoryName: string;
    DepartmentId: number;
    DepartmentName: string;
    PageSize: number;
    PageNumber: number;
    ReceivedCount: number;
    CurrentPage: number;
    SearchString: string;
    IsAllWords: boolean;
}